Support installing Python tools, refactor code
* Support installation of python tools
* Do not install `cargo binstall`
* Support epoch and `v`-prefixed versions for Python
* Use common code to install
* Improve tests

+++ tests.py
@@ -10,13 +10,38 @@ crates_toml = f"""
 [v1]
 "cargo-binstall 1.16.6 (registry+{_crates_index})" = ["cargo-binstall"]
 """
-unknown_tool_ds = 'unknown_tool = {"version" = "1.12.3", "datasource"= "unknown"}'
+unknown_tool = 'unknown_tool = {version = "1.12.3", source = "unknown"}'
+unknown_tool_ver = "unknown_tool"
+install_tool_rs = 'install_tool_rs = {version = "1.1.4", source = "crate"}'
+install_tool_rs_ver = "install_tool_rs@1.1.4"
+install_tool_py = 'install_tool_py = {version = "2.8.9", source = "pypi"}'
+install_tool_py_ver = "install_tool_py==2.8.9"
+dev_tool_py = 'dev_tool_py = {"version" = "1.2.3.dev", source = "pypi"}'
+dev_tool_py_ver = "dev_tool_py==1.2.3.dev"
+epoch_tool_py = 'epoch_tool_py = {"version" = "1!1.2.3", source = "pypi"}'
+epoch_tool_py_ver = "epoch_tool_py==1!1.2.3"
+all_install_tools = "\n".join(
+    (
+        install_tool_rs,
+        install_tool_py,
+        dev_tool_py,
+        epoch_tool_py,
+    )
+)
+all_install_tools_ver = (
+    install_tool_rs_ver,
+    install_tool_py_ver,
+    dev_tool_py_ver,
+    epoch_tool_py_ver,
+)
+
+all_install_tools_unknown = "\n".join((all_install_tools, unknown_tool))
+
 
 SECTION = "section"
-CRATE_DATASOURCE = "crate"
 SECTION_NEVER_INSTALLED = """
 [section-never-installed]
-never_tool = "1.12.3"
+never_tool = {version = "0.0.0", source = "crate"}
 """
 
 
@@ -47,9 +72,9 @@ def setup_tools_toml(
 def execute_command(
     cargo_home: pathlib.Path,
     tool_file: pathlib.Path,
-    cargo_binstall: bool,
-    default_datasource: str | None,
-    force: bool,
+    use_cargo_binstall: bool,
+    use_python_uv: bool,
+    force_install: bool,
     *args: str,
 ):
     """Execute command in debug mode.
@@ -62,14 +87,14 @@ def execute_command(
         "--dry-run",
     ]
 
-    if cargo_binstall:
-        command_args.append("--cargo-binstall")
+    if use_cargo_binstall:
+        command_args.append("--use-cargo-binstall")
 
-    if force:
+    if use_python_uv:
+        command_args.append("--use-python-uv")
+
+    if force_install:
         command_args.append("--force-install")
-
-    if default_datasource is not None:
-        command_args.append(f"--default-datasource={default_datasource}")
 
     # it's better to insert this way to avoid ever-shifting argument index
     command_args.extend(args)
@@ -97,100 +122,30 @@ def execute_command(
 
 @pytest.mark.parametrize("has_crates_toml", (True, False))
 @pytest.mark.parametrize("unknown_section", (True, False))
-@pytest.mark.parametrize("cargo_binstall", (True, False))
-@pytest.mark.parametrize("force", (True, False))
+@pytest.mark.parametrize("use_cargo_binstall", (True, False))
+@pytest.mark.parametrize("use_python_uv", (True, False))
+@pytest.mark.parametrize("force_install", (True, False))
 @pytest.mark.parametrize(
-    "default_datasource,section_data,expected_tools,unsupported_tools",
+    "section_data,expected_tools,unsupported_tools",
     (
-        ("crate", 'install_tool = "1.1.4"', ["install_tool@1.1.4"], []),
-        ("crate", 'install_tool={"version" = "1.1.4"}', ["install_tool@1.1.4"], []),
-        (
-            "crate",
-            'install_tool={"version" = "1.1.4", "datasource" = "crate"}',
-            ["install_tool@1.1.4"],
-            [],
-        ),
-        (
-            "crate",
-            'unknown_tool = {"version" = "1.12.3", "datasource" = "unknown"}',
-            [],
-            ["unknown_tool"],
-        ),
-        (
-            "crate",
-            f'install_tool = "1.1.4"\n{unknown_tool_ds}',
-            ["install_tool@1.1.4"],
-            ["unknown_tool"],
-        ),
-        (
-            "crate",
-            'install_tool = "1.1.4"\nother_tool = {"version" = "1.12.3"}',
-            ["install_tool@1.1.4", "other_tool@1.12.3"],
-            [],
-        ),
-        ("unknown", 'install_tool = "1.1.4"', [], ["install_tool"]),
-        ("unknown", 'install_tool={"version" = "1.1.4"}', [], ["install_tool"]),
-        (
-            "unknown",
-            'install_tool={"version" = "1.1.4", "datasource" = "crate"}',
-            ["install_tool@1.1.4"],
-            [],
-        ),
-        (
-            "unknown",
-            'unknown_tool = {"version" = "1.12.3", "datasource" = "unknown"}',
-            [],
-            ["unknown_tool"],
-        ),
-        (
-            "unknown",
-            f'install_tool = "1.1.4"\n{unknown_tool_ds}',
-            [],
-            ["install_tool", "unknown_tool"],
-        ),
-        (
-            "unknown",
-            'install_tool = "1.1.4"\nother_tool = {"version" = "1.12.3"}',
-            [],
-            ["install_tool", "other_tool"],
-        ),
-        (None, 'install_tool = "1.1.4"', [], ["install_tool"]),
-        (None, 'install_tool={"version" = "1.1.4"}', [], ["install_tool"]),
-        (
-            None,
-            'install_tool={"version" = "1.1.4", "datasource"= "crate"}',
-            ["install_tool@1.1.4"],
-            [],
-        ),
-        (
-            None,
-            unknown_tool_ds,
-            [],
-            ["unknown_tool"],
-        ),
-        (
-            None,
-            f'install_tool = "1.1.4"\n{unknown_tool_ds}',
-            [],
-            ["install_tool", "unknown_tool"],
-        ),
-        (
-            None,
-            'install_tool = "1.1.4"\nother_tool = {"version" = "1.12.3"}',
-            [],
-            ["install_tool", "other_tool"],
-        ),
+        (install_tool_rs, [install_tool_rs_ver], []),
+        (install_tool_py, [install_tool_py_ver], []),
+        (dev_tool_py, [dev_tool_py_ver], []),
+        (epoch_tool_py, [epoch_tool_py_ver], ()),
+        (all_install_tools, all_install_tools_ver, []),
+        (unknown_tool, [], [unknown_tool_ver]),
+        (all_install_tools_unknown, all_install_tools_ver, [unknown_tool_ver]),
     ),
 )
 def test_positive(
     section_data: str,
-    default_datasource: str | None,
     expected_tools: list[str],
     unsupported_tools: list[str],
     has_crates_toml: bool,
     unknown_section: bool,
-    cargo_binstall: bool,
-    force: bool,
+    use_cargo_binstall: bool,
+    use_python_uv: bool,
+    force_install: bool,
     tmpdir: pathlib.Path,
 ):
     expected_tools = list(expected_tools)
@@ -199,24 +154,33 @@ def test_positive(
     unsupported_tool_idx = 0
     cargo_home = tmpdir / ".cargo"
     tool_file = tmpdir / "tools.toml"
+
     setup_cargo_home(cargo_home, has_crates_toml)
     setup_tools_toml(tool_file, section_data, unknown_section, raw=False)
 
     (code, stdout, stderr) = execute_command(
         cargo_home,
         tool_file,
-        cargo_binstall,
-        default_datasource,
-        force,
+        use_cargo_binstall,
+        use_python_uv,
+        force_install,
     )
     assert len(stdout) == 0
     assert code == 0
+
+    warning_python_force_flag = False
 
     for line in stderr.splitlines():
         assert "[ERROR]" not in line
         if "[DEBUG]" in line or "[INFO]" in line:
             continue
 
+        warning_python_force_flag = warning_python_force_flag or line.endswith(
+            "[WARNING] Force install flag is not yet supported for Python packages"
+        )
+
+        if warning_python_force_flag:
+            continue
         install_tool = re.sub(
             "^.* \\[WARNING] Running install command: '(.*)'$", "\\1", line
         )
@@ -235,27 +199,70 @@ def test_positive(
             expected_tool = expected_tools[expected_tool_idx]
             expected_tool_idx += 1
 
-            assert cmd[0] == "cargo"
-            if cargo_binstall:
-                assert cmd[1] == "binstall"
-            else:
-                assert cmd[1] == "install"
-
-            idx = 2
-            if cargo_binstall:
-                assert cmd[idx] == "--no-confirm"
-                idx += 1
-
-            if force:
-                assert cmd[idx] == "--force"
-                idx += 1
-
-            assert cmd[idx] == expected_tool
+            match cmd[0]:
+                case "cargo":
+                    check_cargo_install(
+                        cmd, expected_tool, use_cargo_binstall, force_install
+                    )
+                case "uv" | "pip":
+                    check_python_install(
+                        cmd,
+                        expected_tool,
+                        use_python_uv,
+                        force_install,
+                        warning_python_force_flag,
+                    )
+                case _:
+                    assert False, f"{cmd[0]} is unsupported: {install_tool!r}"
 
         if unsupported:
             assert len(unsupported_tools) > unsupported_tool_idx, unsupported_tools
-            assert unsupported == unsupported_tools[unsupported_tool_idx]
+            assert unsupported == unsupported_tools[unsupported_tool_idx], line
             unsupported_tool_idx += 1
+
+
+def check_cargo_install(
+    cmd: list[str], expected_tool: str, use_cargo_binstall: bool, force_install: bool
+):
+    """Checks tool installation using cargo"""
+    assert cmd[0] == "cargo"
+    if use_cargo_binstall:
+        assert cmd[1] == "binstall"
+    else:
+        assert cmd[1] == "install"
+
+    idx = 2
+    if use_cargo_binstall:
+        assert cmd[idx] == "--no-confirm"
+        idx += 1
+
+    if force_install:
+        assert cmd[idx] == "--force"
+        idx += 1
+
+    assert cmd[idx] == expected_tool
+
+    assert len(cmd) == idx + 1
+
+
+def check_python_install(
+    cmd: list[str],
+    expected_tool: str,
+    use_python_uv: bool,
+    force_install: bool,
+    warning_python_force_flag: bool,
+):
+    idx = 0
+    if use_python_uv:
+        assert cmd[0] == "uv"
+        idx += 1
+
+    assert cmd[idx] == "pip"
+    assert cmd[idx + 1] == "install"
+    assert cmd[idx + 2] == expected_tool
+
+    if force_install:
+        assert warning_python_force_flag, "Warning hasn't been met"
 
 
 @pytest.mark.parametrize(
@@ -266,23 +273,14 @@ def test_positive(
         ('_invalid_tool = "1.2.3"', "Tool name is unexpected: '_invalid_tool'"),
         ('"invalid_tool*" = "1.2.3"', "Tool name is unexpected: 'invalid_tool*'"),
         # version name is unexpected or unexpected type
-        ("invalid_tool = true", "invalid_tool: Value must be a string or dict"),
-        ("invalid_tool = false", "invalid_tool: Value must be a string or dict"),
-        ("invalid_tool = 123", "invalid_tool: Value must be a string or dict"),
-        (
-            'invalid_tool = "Wow"',
-            "invalid_tool: Version doesn't match semver scheme: 'Wow'",
-        ),
-        # Semver suffix
-        (
-            'invalid_tool = "1.2.3.dev"',
-            "invalid_tool: Version doesn't match semver scheme: '1.2.3.dev'",
-        ),
-        # Epoch prefix (Python specific)
-        (
-            'invalid_tool = "20!1.2.3"',
-            "invalid_tool: Version doesn't match semver scheme: '20!1.2.3'",
-        ),
+        ("invalid_tool = true", "invalid_tool: Value must be a dict"),
+        ("invalid_tool = false", "invalid_tool: Value must be a dict"),
+        ("invalid_tool = 123", "invalid_tool: Value must be a dict"),
+        ('invalid_tool = "Wow"', "invalid_tool: Value must be a dict"),
+        ('invalid_tool = "1.2.3"', "invalid_tool: Value must be a dict"),
+        ('invalid_tool = "1.2.3.dev"', "invalid_tool: Value must be a dict"),
+        ('invalid_tool = "20!1.2.3"', "invalid_tool: Value must be a dict"),
+        # dversion is mandatory
         ("invalid_tool = {}", "invalid_tool: Version is mandatory"),
         ('invalid_tool = {"key" = "value"}', "invalid_tool: Version is mandatory"),
         ('invalid_tool = {"version" = {}}', "invalid_tool: Version must be a string"),
@@ -295,14 +293,6 @@ def test_positive(
         (
             'invalid_tool = {"version" = "Wow"}',
             "invalid_tool: Version doesn't match semver scheme: 'Wow'",
-        ),
-        (
-            'invalid_tool = {"version" = "1.2.3.dev"}',
-            "invalid_tool: Version doesn't match semver scheme: '1.2.3.dev'",
-        ),
-        (
-            'invalid_tool = {"version" = "20!1.2.3"}',
-            "invalid_tool: Version doesn't match semver scheme: '20!1.2.3'",
         ),
     ),
 )
@@ -319,9 +309,9 @@ def test_negative(
     (code, stdout, stderr) = execute_command(
         cargo_home,
         tool_file,
-        True,
-        CRATE_DATASOURCE,
-        True,
+        use_cargo_binstall=True,
+        use_python_uv=True,
+        force_install=True,
     )
     assert len(stdout) == 0
     assert code != 0
